simplify ContactLinks, add descrip
